Updated so it doesn't need run_makeroom at all, since that was removing everything.

+++ makeroom.sh
@@ -60,9 +60,9 @@ TOOL_ROOT=${TOOL_ROOT//[-.]/_}
 ############## Check if the tools are in the PATH ###################
 if [ -z "$UPPMAX_ROOT" ]
 then
-    printf "No path to this tool found. Make sure you have https://github.com/UPPMAX/install-methods cloned and in your \$PATH\n\n" >&2 && exit 1
+    printf "No path to this tool found. Make sure you have https://github.com/UPPMAX/install-methods cloned and findable, e.g. in your \$PATH\n\n" >&2 && exit 1
 else
-    printf "Found tools at %s\n\n" "$UPPMAX_ROOT"
+    printf "Found tools at %s\n\n" "$UPPMAX_ROOT" >&2
 fi
 
 ############################
@@ -123,7 +123,7 @@ do
             FIXFLAG+=($OPTARG)
             USERGROUP=$OPTARG
             ;;
-        P) FIXFLAG+=(-G)
+        P) FIXFLAG+=(-P)
             FIXFLAG+=($OPTARG)
             USERPERMISSIONS="$OPTARG"
             ;;
@@ -378,10 +378,17 @@ NEWS6="$LICENSE"
 
 ################### If resuming, now we exit #####################
 if [ $MODE == "RESUME" ] ; then
-    printf "export TOOL=%s VERSION=%s TOOLDIR=%s VERSIONDIR=%s PREFIX=%s COMMONDIR=%s SRCDIR=%s \nexport NEWS=\"%s\n%s\n%s\n%s\n%s\n%s\"" "$TOOL" "$VERSION" "$TOOLDIR" "$VERSIONDIR" "i$PREFIX" "$COMMONDIR" "$SRCDIR" "${NEWS1}" "${NEWS2}" "${NEWS3}" "${NEWS4}" "${NEWS5}" "${NEWS6}" > $TMPFILE
+    if [ -f $SOURCEMEFILE ]; then
+        printf "Run this to get your variables:\n" 1>&2
+        printf "source %s\n" "$SOURCEMEFILE" 1>&2
+    else
+        printf "export TOOL=%s VERSION=%s TOOLDIR=%s VERSIONDIR=%s PREFIX=%s COMMONDIR=%s SRCDIR=%s \nexport NEWS=\"%s\n%s\n%s\n%s\n%s\n%s\"" "$TOOL" "$VERSION" "$TOOLDIR" "$VERSIONDIR" "i$PREFIX" "$COMMONDIR" "$SRCDIR" "${NEWS1}" "${NEWS2}" "${NEWS3}" "${NEWS4}" "${NEWS5}" "${NEWS6}" > $TMPFILE
+        printf "No previous install found, but you can run this to get your variables:\n" 1>&2
+        printf "source %s\n" "$TMPFILE" 1>&2
+    fi
     ## echo for run_makeroom
-    echo $TMPFILE
-    exit 0
+    #echo $TMPFILE
+    #exit 0
 fi
 ####################### REMOVE ############################################
 
@@ -456,7 +463,7 @@ RMVTMP
     printf "export TOOL='' VERSION='' TOOLDIR='' VERSIONDIR='' PREFIX='' COMMONDIR='' SRCDIR='' NEWS=''" > $TMPFILE
     chmod +x $TMPFILE
     ## echo for run_makeroom
-    printf "%s\n" $TMPFILE
+    #printf "%s\n" $TMPFILE
     chmod +x $REMOVEFILE
     exit 0;
 fi
@@ -580,7 +587,7 @@ prepend-path    PYTHONPATH          \\\$modroot/lib/python2.7/site-packages
 prepend-path    MANPATH             \\\$modroot/share/man
 prepend-path    CPATH               \\\$modroot/include
 prepend-path    CPLUS_INCLUDE_PATH  \\\$modroot/include
-setenv          $TOOL_ROOT      \\\$modroot
+setenv          $TOOL_ROOT          \\\$modroot
 #### If you need the user's home ###########
 #setenv          HOME                \\\$::env(HOME)
 
@@ -693,13 +700,17 @@ echo "$NEWS6" 1>&2
 
 umask \$PREUMASK
 
-mv $PWD/$SCRIPTFILE /sw/$CATEGORY/${TOOL}/
+mv $PWD/$SCRIPTFILE ${TOOLDIR}/
 printf "export TOOL=%s VERSION=%s TOOLDIR=%s VERSIONDIR=%s PREFIX=%s COMMONDIR=%s SRCDIR=%s \nexport NEWS=\"%s\n%s\n%s\n%s\n%s\n%s\"" "$TOOL" "$VERSION" "$TOOLDIR" "$VERSIONDIR" "/sw/$CATEGORY/$TOOL/$VERSION/\\\$CLUSTER" "$COMMONDIR" "$SRCDIR" "${NEWS1}" "${NEWS2}" "${NEWS3}" "${NEWS4}" "${NEWS5}" "${NEWS6}" > $SOURCEMEFILE
+printf "Run this to set your variables and go to \$TOOL:\n\n"
+printf "source %s && cd \$TOOLDIR\n" $SOURCEMEFILE
 TMP
 
 ################# End of installation makeroom script #########################
 
 printf "export TOOL=%s VERSION=%s TOOLDIR=%s VERSIONDIR=%s PREFIX=%s COMMONDIR=%s SRCDIR=%s \nexport NEWS=\"%s\n%s\n%s\n%s\n%s\n%s\"" "$TOOL" "$VERSION" "$TOOLDIR" "$VERSIONDIR" "$PREFIX" "$COMMONDIR" "$SRCDIR" "${NEWS1}" "${NEWS2}" "${NEWS3}" "${NEWS4}" "${NEWS5}" "${NEWS6}" > $TMPFILE
+printf "No files yet created in the file tree, but you can run this to get your variables set:\n" 1>&2
+printf "source %s\n" "$TMPFILE" 1>&2
     ## echo for run_makeroom
-echo $TMPFILE
+    #echo $TMPFILE
 chmod +x $SCRIPTFILE
